deleting files and adding, and adding to index for better imports
