+++ apps/judicial-system/web/src/components/ReactTable/DefendantInfo/DefendantInfo.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css'
+
+export const blockColumn = style({
+  display: 'block',
+  overflow: 'hidden',
+  textOverflow: 'ellipsis',
+})
