+++ apps/judicial-system/web/src/components/ReactTable/DefendantInfo/DefendantInfo.tsx
@@ -1,0 +1,40 @@
+import React from 'react'
+
+import { Box, Text } from '@island.is/island-ui/core'
+import { formatDOB } from '@island.is/judicial-system/formatters'
+import { Defendant } from '@island.is/judicial-system/types'
+
+import * as styles from './DefendantInfo.css'
+
+interface Props {
+  defendants?: Defendant[]
+}
+
+const DefendantInfo: React.FC<Props> = ({ defendants }) => {
+  return defendants && defendants.length > 0 ? (
+    <>
+      <Text>
+        <Box component="span" className={styles.blockColumn}>
+          {defendants[0].name ?? '-'}
+        </Box>
+      </Text>
+      {defendants.length === 1 ? (
+        (!defendants[0].noNationalId || defendants[0].nationalId) && (
+          <Text>
+            <Text as="span" variant="small" color="dark400">
+              {formatDOB(defendants[0].nationalId, defendants[0].noNationalId)}
+            </Text>
+          </Text>
+        )
+      ) : (
+        <Text as="span" variant="small" color="dark400">
+          {`+ ${defendants.length - 1}`}
+        </Text>
+      )}
+    </>
+  ) : (
+    <Text>-</Text>
+  )
+}
+
+export default DefendantInfo
