+++ apps/judicial-system/web/src/components/Table/CategoryCard/CategoryCard.css.ts
