+++ apps/judicial-system/web/src/components/Table/PastCasesTable/PastCases.spec.ts
