+++ apps/judicial-system/web/src/components/Table/Table.css.ts
@@ -6,8 +6,69 @@ export const header = style({
 })
 
 export const table = style({
+  borderSpacing: 0,
   borderCollapse: 'collapse',
+  overflow: 'hidden',
+
+  // Needed for Safari.
   width: '100%',
+})
+
+export const thead = style({
+  background: theme.color.blue100,
+  textAlign: 'left',
+})
+
+export const thButton = style({
+  outline: 'none',
+
+  ':active': {
+    color: theme.color.dark400,
+  },
+})
+
+export const tableRowContainer = style({
+  borderBottom: `1px solid ${theme.color.blue200}`,
+  cursor: 'pointer',
+})
+
+export const th = style({
+  padding: `${theme.spacing[2]}px ${theme.spacing[3]}px`,
+})
+
+export const largeColumn = style({
+  '@media': {
+    [`screen and (min-width: ${theme.breakpoints.xl}px)`]: {
+      // The width needed to make sure a 33 character name doesn't wrap
+      maxWidth: 334,
+      whiteSpace: 'nowrap',
+    },
+  },
+})
+export const blockColumn = style({
+  display: 'block',
+  overflow: 'hidden',
+  textOverflow: 'ellipsis',
+})
+
+export const deleteButtonContainer = style({
+  maxWidth: '0',
+  height: '100%',
+  padding: 0,
+  transform: 'translate3d(2px, 0px, 0px)',
+})
+
+export const td = style({
+  selectors: {
+    [`&:not(${deleteButtonContainer})`]: {
+      padding: `${theme.spacing[2]}px ${theme.spacing[3]}px`,
+    },
+    '&.secondLast': {
+      marginLeft: 'auto',
+      height: '100%',
+      padding: 0,
+    },
+  },
 })
 
 globalStyle(`${table} td, th`, {
@@ -28,4 +89,35 @@ export const expandLabel = style({
 
 export const row = style({
   cursor: 'pointer',
+})
+
+export const sortIcon = style({
+  opacity: 0.4,
+  transition: 'opacity .2s ease-in-out',
+
+  selectors: {
+    [`${thButton}:hover &`]: {
+      opacity: 1,
+    },
+  },
+})
+
+export const sortCreatedAsc = style({
+  opacity: 1,
+  transform: 'rotate(180deg)',
+})
+
+export const sortCreatedDes = style({
+  opacity: 1,
+  transform: 'rotate(0deg)',
+})
+
+export const sortAccusedNameAsc = style({
+  opacity: 1,
+  transform: 'rotate(0deg)',
+})
+
+export const sortAccusedNameDes = style({
+  opacity: 1,
+  transform: 'rotate(180deg)',
 })
