+++ apps/judicial-system/web/src/components/Table/TableHeaderText/TableHeaderText.tsx
