+++ apps/judicial-system/web/src/components/Table/TableSkeleton/TableSkeleton.tsx
