+++ apps/judicial-system/web/src/components/Table/index.ts
@@ -1,0 +1,7 @@
+export { default as CategoryCard } from './CategoryCard/CategoryCard'
+export { default as ColumnCaseType } from './ColumnCaseType/ColumnCaseType'
+export { default as CourtCaseNumber } from './CourtCaseNumber/CourtCaseNumber'
+export { default as DefendantInfo } from './DefendantInfo/DefendantInfo'
+export { default as SortButton } from './SortButton/SortButton'
+export { default as TableContainer } from './TableContainer/TableContainer'
+export { default as TableHeaderText } from './TableHeaderText/TableHeaderText'
