+++ apps/judicial-system/web/src/components/index.ts
@@ -58,13 +58,17 @@ export { default as SelectCourtOfficials } from './SelectCourtOfficials/SelectCo
 export { default as ServiceInterruptionBanner } from './ServiceInterruptionBanner/ServiceInterruptionBanner'
 export { default as SignedDocument } from './SignedDocument/SignedDocument'
 export { default as TagAppealState } from './TagAppealState/TagAppealState'
+export { default as TagCaseState } from './TagCaseState/TagCaseState'
+
+export { default as AppealCasesTable } from './Table/AppealCasesTable/AppealCasesTable'
+export { default as PastCasesTable } from './Table/PastCasesTable/PastCasesTable'
 
 export {
   SigningModal,
   useRequestRulingSignature,
 } from './SigningModal/SigningModal'
 export { default as Skeleton } from './Skeleton/Skeleton'
-export { default as Table } from './Table/Table'
+export { default as Table } from './ReactTable/Table'
 export { default as TimeInputField } from './TimeInputField/TimeInputField'
 export { UserProvider, UserContext } from './UserProvider/UserProvider'
 export {
